Move Type definitions to seperate file
